python/requirementsCLI.txt
numpy==1.21.6
scipy==1.7.3
pyproj==3.5.0
biopython==1.81
geojson==3.0.1
svgwrite==1.4.3
python-datauri==2.1.0
requests==2.31.0